api-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.8.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
asyncio==3.4.3
httpx==0.25.2